+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Add CHANGELOG.md
   - Add name/email to source file
   - Display time elapsed when playback ended
-  
+  - Add experimental desktop notifications with libnotify
+
 ### June 4th, 2019
 - First Release, v1.0